Move the declarations away from `build.cc`
First step towards actually splitting the implementation
